requirements.txt
xmltodict
jinja2
catilo
requests
rcssmin
rjsmin
markdown
Pillow
beautifulsoup4
lxml